requirements.txt
streamlit==1.35.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.3.0
langdetect==1.0.9
python-dotenv==1.0.1
